introduce modification_mapper.py
